refactor!: move tket2 extension definition to `extension.rs`
